requirements.txt
requests
beautifulsoup4
langchain
sentence-transformers
faiss-cpu
transformers
torch
bitsandbytes 
accelerate